requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.29,<1.0
openai>=1.12,<2.0
httpx>=0.27,<1.0
pyyaml>=6.0,<7.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
